Microgrammar tests no longer abstract. Simplifies examining test failures
